sync

+++ engine/Shopware/Controllers/Frontend/Payment.php
@@ -29,7 +29,7 @@ use Shopware\Components\Random;
 /**
  * Shopware Payment Controller
  *
- * @category  Shopware
+ * @category Shopware
  *
  * @copyright Copyright (c) shopware AG (http://www.shopware.de)
  */
@@ -38,7 +38,7 @@ abstract class Shopware_Controllers_Frontend_Payment extends Enlight_Controller_
     /**
      * Returns the current payment short name.
      *
-     * @return string
+     * @return string|null
      */
     public function getPaymentShortName()
     {
@@ -78,7 +78,7 @@ abstract class Shopware_Controllers_Frontend_Payment extends Enlight_Controller_
      * @param int    $paymentStatusId
      * @param bool   $sendStatusMail
      *
-     * @return int
+     * @return int|false
      */
     public function saveOrder($transactionId, $paymentUniqueId, $paymentStatusId = null, $sendStatusMail = false)
     {
@@ -141,7 +141,7 @@ abstract class Shopware_Controllers_Frontend_Payment extends Enlight_Controller_
             WHERE transactionID=? AND temporaryID=?
             AND status!=-1
         ';
-        $orderId = Shopware()->Db()->fetchOne($sql, [
+        $orderId = (int) Shopware()->Db()->fetchOne($sql, [
                 $transactionId,
                 $paymentUniqueId,
             ]);
@@ -168,7 +168,7 @@ abstract class Shopware_Controllers_Frontend_Payment extends Enlight_Controller_
     /**
      * Returns shipment amount as float
      *
-     * @return float
+     * @return string
      */
     public function getShipment()
     {
@@ -184,7 +184,7 @@ abstract class Shopware_Controllers_Frontend_Payment extends Enlight_Controller_
     /**
      * Returns the full user data as array.
      *
-     * @return array
+     * @return array|null
      */
     public function getUser()
     {
@@ -198,7 +198,7 @@ abstract class Shopware_Controllers_Frontend_Payment extends Enlight_Controller_
     /**
      * Returns the full basket data as array.
      *
-     * @return array
+     * @return array|null
      */
     public function getBasket()
     {
@@ -273,8 +273,7 @@ abstract class Shopware_Controllers_Frontend_Payment extends Enlight_Controller_
     }
 
     /**
-     * @param string      $signature
-     * @param ArrayObject $basket
+     * @param string $signature
      *
      * @throws RuntimeException if signature does not match with provided basket
      */
